fix(video): make video controls work in all browsers
The controls were previously not working in Safari, e.g. pausing and forwarding would just restart the video. I couldn't exactly pin down the problem, but it was caused by the script that `innerHTML`ed the video element into the DOM. Displaying the native video element by default might not be as visually appealing as the previous solution, but I would argue that usability takes precedence here.